refactor(search applications): rename search preview folder
