+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/engine_view.tsx
@@ -29,10 +29,10 @@ import { DeleteSearchApplicationModal } from '../search_applications/delete_sear
 
 import { EngineConnect } from './engine_connect/engine_connect';
 import { EngineError } from './engine_error';
-import { EngineSearchPreview } from './engine_search_preview/engine_search_preview';
 import { EngineViewLogic } from './engine_view_logic';
 import { EngineHeaderDocsAction } from './header_docs_action';
 import { SearchApplicationContent } from './search_application_content';
+import { EngineSearchPreview } from './search_preview/engine_search_preview';
 
 export const EngineView: React.FC = () => {
   const { fetchEngine, closeDeleteEngineModal } = useActions(EngineViewLogic);
