+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/search_preview/convert_results.ts
