+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/search_preview/document_context.test.tsx
